requirements.txt
aiogram==3.15.0
python-pptx==1.0.2
pillow==11.0.0
openai==1.57.4
python-dotenv==1.0.1
